# CI: speed up FW unit tests