@sshane 2023-08-12
oh, forgot I have https://github.com/commaai/openpilot/pull/28435 🤦‍♂️